tests() added test for vec3.fromValues()
